add subnet suit struct and methods
